+++ Avancerad.NET-Projekt/Controllers/CustomerController.cs
@@ -6,6 +6,13 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Identity;
+using System.Security.Claims;
+using Microsoft.IdentityModel.Tokens;
+using System.IdentityModel.Tokens.Jwt;
+using System.Text;
+using System.Numerics;
 
 namespace Avancerad.NET_Projekt.Controllers
 {
@@ -13,16 +20,18 @@ namespace Avancerad.NET_Projekt.Controllers
     [ApiController]
     public class CustomerController : ControllerBase
     {
+        private readonly UserManager<IdentityUser> _userManager;
         private readonly ICustomerRepo _customerRepo;
         private readonly IMapper _mapper;
 
-        public CustomerController(ICustomerRepo customerRepo, IMapper mapper)
+        public CustomerController(ICustomerRepo customerRepo, IMapper mapper, UserManager<IdentityUser> userManager)
         {
             _customerRepo = customerRepo;
             _mapper = mapper;
+            _userManager = userManager;
         }
 
-        [HttpGet, Authorize]
+        [HttpGet, Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Policy = "AdminCompanyPolicy")]
         public async Task<ActionResult<IEnumerable<CustomerDTO>>> GetAllCustomers()
         {
             try
@@ -39,7 +48,7 @@ namespace Avancerad.NET_Projekt.Controllers
             }
         }
 
-        [HttpGet("appointments/{id:int}")]
+        [HttpGet("appointments/{id:int}"), Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Policy = "AdminCompanyPolicy")]
         public async Task<ActionResult<object>> GetCustomerAppointments(int id)
         {
             try
@@ -52,13 +61,7 @@ namespace Avancerad.NET_Projekt.Controllers
                 }
 
                 // Filter out appointments where IsDeleted is false
-                var appointments = customer.Appointments.Where(a => !a.IsDeleted).ToList();
-
-                // Check if the customer has any non-deleted appointments
-                if (appointments.Count == 0)
-                {
-                    return NotFound($"Customer with ID {id} does not have any active appointments.");
-                }
+                var appointments = customer.Appointments/*.Where(a => !a.IsDeleted)*/.ToList();
 
                 var customerDTO = _mapper.Map<CustomerDTO>(customer);
                 var appointmentsDTO = _mapper.Map<List<AppointmentDTO>>(appointments);
@@ -79,7 +82,7 @@ namespace Avancerad.NET_Projekt.Controllers
             }
         }
 
-        [HttpGet("{id:int}")]
+        [HttpGet("{id:int}"), Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Policy = "AdminCompanyPolicy")]
         public async Task<ActionResult<CustomerDTO>> GetCustomer(int id)
         {
             try
@@ -102,37 +105,47 @@ namespace Avancerad.NET_Projekt.Controllers
         }
 
         [HttpPost]
-        public async Task<ActionResult<CustomerDTO>> CreateCustomer(CustomerDTO newCustomerDTO)
+        [Route("account/create/customer")]
+        public async Task<IActionResult> CreateCompany(string email, string password, string firstName, string lastName, string phone)
         {
-            try
+            IdentityUser User = await _userManager.FindByEmailAsync(email);
+            if (User != null)
+                return BadRequest(false);
+
+            IdentityUser user = new()
             {
-                if (newCustomerDTO == null)
-                {
-                    return BadRequest();
-                }
+                UserName = email,
+                PasswordHash = password,
+                Email = email,
+            };
 
-                // Map CustomerDTO to Customer
-                var newCustomer = _mapper.Map<Customer>(newCustomerDTO);
+            IdentityResult result = await _userManager.CreateAsync(user, password);
 
-                // Create the customer using the service
-                var createdCustomer = await _customerRepo.CreateAsync(newCustomer);
+            if (!result.Succeeded)
+                return BadRequest(false);
 
-                // Map the created Customer back to CustomerDTO
-                var createdCustomerDTO = _mapper.Map<CustomerDTO>(createdCustomer);
+            Claim[] userClaims =
+                [
+                    new Claim(ClaimTypes.Email, email),
+            new Claim(ClaimTypes.Role, "company")
+                ];
+            await _userManager.AddClaimsAsync(user, userClaims);
 
-                return CreatedAtAction(nameof(GetCustomer),
-                    new { id = createdCustomer.CustomerID },
-                    createdCustomerDTO);
-            }
-            catch (Exception)
+            var customer = new Customer
             {
-                return StatusCode(StatusCodes.Status500InternalServerError,
-                    "Error creating data in the database.");
-            }
+                IdentityUser = user,
+                FirstName = firstName,
+                LastName = lastName,
+                Phone = phone,
+                Email = email,
+            };
+            await _customerRepo.CreateAsync(customer);
+
+            return Ok(true);
         }
 
-        [HttpDelete("{id:int}")]
-        public async Task<ActionResult<CustomerDTO>> DeleteCustomer(int id)
+        [HttpDelete("{id:int}"), Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Policy = "AdminCompanyCustomerPolicy")]
+        public async Task<ActionResult<CustomerDTO>> DeleteCustomer(int id, [FromServices] UserManager<IdentityUser> userManager)
         {
             try
             {
@@ -145,24 +158,52 @@ namespace Avancerad.NET_Projekt.Controllers
                     return NotFound($"Customer with ID {id} does not exist and therefore cannot be deleted.");
                 }
 
+                var email = User.FindFirstValue(ClaimTypes.Email);
+                var loggedInUser = await userManager.FindByEmailAsync(email);
+
+                // Check if the logged-in user exists
+                if (loggedInUser == null)
+                {
+                    return StatusCode(StatusCodes.Status500InternalServerError,
+                        "Error retrieving the logged-in user.");
+                }
+
+                // Check if the logged-in user is trying to delete their own account or if they have the admin or company role
+                if (loggedInUser.Id != customerToDelete.IdentityUserId && User != null && (!User.IsInRole("admin") && !User.IsInRole("company")))
+                {
+                    return Forbid();
+                }
+
                 // Map the customer to CustomerDTO
                 var customerDTOToReturn = _mapper.Map<CustomerDTO>(customerToDelete);
 
                 // Perform the deletion
                 await _customerRepo.DeleteAsync(id);
 
+                // Delete the associated IdentityUser
+                var identityUserToDelete = await userManager.FindByIdAsync(customerToDelete.IdentityUserId);
+                if (identityUserToDelete != null)
+                {
+                    var identityResult = await userManager.DeleteAsync(identityUserToDelete);
+                    if (!identityResult.Succeeded)
+                    {
+                        return StatusCode(StatusCodes.Status500InternalServerError,
+                            "Error deleting associated IdentityUser from the database.");
+                    }
+                }
+
                 // Return the CustomerDTO with a status code of OK (HTTP 200)
                 return Ok(customerDTOToReturn);
             }
-            catch (Exception)
+            catch (Exception ex)
             {
                 return StatusCode(StatusCodes.Status500InternalServerError,
-                    "Error deleting data from the database.");
+                    $"Error deleting data from the database. {ex.Message}");
             }
         }
 
-        [HttpPut("{id:int}")]
-        public async Task<ActionResult> UpdateCustomer(int id, CustomerDTO customerDTO)
+        [HttpPut("{id:int}"), Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Policy = "AdminCompanyCustomerPolicy")]
+        public async Task<ActionResult> UpdateCustomer(int id, CustomerUpdateDTO customerDTO, [FromServices] UserManager<IdentityUser> userManager)
         {
             try
             {
@@ -179,6 +220,22 @@ namespace Avancerad.NET_Projekt.Controllers
                     return NotFound($"Customer with ID {id} does not exist.");
                 }
 
+                var email = User.FindFirstValue(ClaimTypes.Email);
+                var loggedInUser = await userManager.FindByEmailAsync(email);
+
+                // Check if the logged-in user exists
+                if (loggedInUser == null)
+                {
+                    return StatusCode(StatusCodes.Status500InternalServerError,
+                        "Error retrieving the logged-in user.");
+                }
+
+                // Check if the logged-in user is trying to update their own account or if they have the admin or company role
+                if (loggedInUser.Id != existingCustomer.IdentityUserId && User != null && (!User.IsInRole("admin") && !User.IsInRole("company")))
+                {
+                    return Forbid();
+                }
+
                 // Map the CustomerDTO to Customer and set the ID
                 var updatedCustomer = _mapper.Map<Customer>(customerDTO);
                 updatedCustomer.CustomerID = id;
@@ -187,14 +244,15 @@ namespace Avancerad.NET_Projekt.Controllers
                 await _customerRepo.UpdateAsync(updatedCustomer);
                 return NoContent(); // Return 204 No Content after successful update
             }
-            catch (Exception)
+            catch (Exception ex)
             {
                 return StatusCode(StatusCodes.Status500InternalServerError,
-                    "Error updating data in the database.");
+                    $"Error updating data in the database. {ex.Message}");
             }
         }
 
-        [HttpGet("customers/{customerId:int}/appointments/week/{year:int}/{weekNumber:int}")]
+        [HttpGet("customers/{customerId:int}/appointments/week/{year:int}/{weekNumber:int}"), 
+            Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Policy = "AdminCompanyPolicy")]
         public async Task<ActionResult<int>> GetCustomerAppointmentsInWeek(int customerId, int year, int weekNumber)
         {
             try
